cli/sql: extract the dispatch code of the shell into a new file
This ensures there are no cobra dependencies remaining inside the
interactive shell.

Release note: None
